Create a new MapModel to store map-related info
