+++ common/src/main/java/com/wynntils/features/user/overlays/map/MapFeature.java
@@ -4,15 +4,18 @@
  */
 package com.wynntils.features.user.overlays.map;
 
+import com.google.common.collect.ImmutableList;
 import com.wynntils.core.config.Config;
 import com.wynntils.core.features.UserFeature;
 import com.wynntils.core.features.properties.FeatureCategory;
 import com.wynntils.core.features.properties.FeatureInfo;
 import com.wynntils.core.features.properties.RegisterKeyBind;
 import com.wynntils.core.keybinds.KeyBind;
+import com.wynntils.core.managers.Model;
 import com.wynntils.gui.screens.maps.MainMapScreen;
 import com.wynntils.mc.objects.CustomColor;
 import com.wynntils.mc.utils.McUtils;
+import com.wynntils.wynn.model.map.MapModel;
 import org.lwjgl.glfw.GLFW;
 
 @FeatureInfo(category = FeatureCategory.MAP)
@@ -36,4 +39,10 @@ public class MapFeature extends UserFeature {
 
         McUtils.mc().setScreen(new MainMapScreen());
     });
+
+    @Override
+    protected void onInit(ImmutableList.Builder<Condition> conditions, ImmutableList.Builder<Class<? extends Model>> dependencies) {
+        super.onInit(conditions, dependencies);
+        dependencies.add(MapModel.class);
+    }
 }
