+++ common/src/main/java/com/wynntils/gui/render/MapRenderer.java
@@ -12,7 +12,7 @@ import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.Tesselator;
 import com.mojang.blaze3d.vertex.VertexFormat;
 import com.mojang.math.Matrix4f;
-import com.wynntils.core.webapi.profiles.MapProfile;
+import com.wynntils.wynn.model.map.MapProfile;
 import com.wynntils.features.user.overlays.map.PointerType;
 import com.wynntils.mc.objects.CustomColor;
 import com.wynntils.mc.utils.McUtils;
