+++ common/src/main/java/com/wynntils/gui/screens/maps/MainMapScreen.java
@@ -7,8 +7,7 @@ package com.wynntils.gui.screens.maps;
 import com.mojang.blaze3d.platform.InputConstants;
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.PoseStack;
-import com.wynntils.core.webapi.WebManager;
-import com.wynntils.core.webapi.profiles.MapProfile;
+import com.wynntils.wynn.model.map.MapProfile;
 import com.wynntils.features.user.overlays.map.MapFeature;
 import com.wynntils.gui.render.MapRenderer;
 import com.wynntils.gui.render.RenderUtils;
@@ -19,6 +18,7 @@ import com.wynntils.mc.utils.McUtils;
 import com.wynntils.utils.KeyboardUtils;
 import com.wynntils.utils.MathUtils;
 import com.wynntils.wynn.model.CompassModel;
+import com.wynntils.wynn.model.map.MapModel;
 import net.minecraft.client.KeyMapping;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.resources.sounds.SimpleSoundInstance;
@@ -126,9 +126,9 @@ public class MainMapScreen extends Screen {
 
         renderBackground(poseStack);
 
-        if (!WebManager.isMapLoaded()) return;
+        if (!MapModel.isMapLoaded()) return;
 
-        MapProfile map = WebManager.getMaps().get(0);
+        MapProfile map = MapModel.getMaps().get(0);
         float textureX = map.getTextureXPosition(mapCenterX);
         float textureZ = map.getTextureZPosition(mapCenterZ);
